refactor: update code according to rewiew
