+++ src/modules/TextBookPage/components/WordCard/components/ButtonsBlock/index.ts
