+++ src/modules/TextBookPage/components/WordCard/components/ButtonsBlock/styled.ts
