+++ src/modules/TextBookPage/components/WordCard/components/SentencesBlock/SentencesBlock.tsx
