+++ src/modules/TextBookPage/components/WordCard/components/SentencesBlock/index.ts
