+++ src/modules/TextBookPage/components/WordCard/components/SentencesBlock/styled.ts
