+++ src/modules/TextBookPage/components/WordCard/components/TopPart/index.ts
