+++ src/modules/TextBookPage/components/WordCard/components/TopPart/styled.ts
