+++ src/modules/TextBookPage/components/WordCard/components/index.ts
@@ -1,0 +1,3 @@
+export { ButtonsBlock } from './ButtonsBlock';
+export { SentencesBlock } from './SentencesBlock';
+export { TopPart } from './TopPart';
