+++ src/modules/TextBookPage/components/WordsList/WordList.tsx
@@ -3,8 +3,8 @@ import { Word } from 'types';
 import { Paper } from '@material-ui/core';
 import { useSelector } from 'react-redux';
 import { selectGroup } from 'modules/TextBookPage/selectors';
+import { LEVEL_COLORS } from 'appConstants/colors';
 import { WordCard } from '../WordCard';
-import { LEVEL_COLORS } from '../../../../appConstants/colors';
 import { WordListStyled } from './styled';
 
 type WordListProps = {
@@ -20,8 +20,8 @@ export const WordList: FC<WordListProps> = ({ words }) => {
           <WordCard
             word={word}
             colorGroup={LEVEL_COLORS[group]}
-            countIntoRedBlock={5}
-            countIntoGreenBlock={10}
+            successCount={5}
+            errorCount={10}
             isTranslate={true}
             isButtons={true}
           />
