+++ tsconfig.json
@@ -1,11 +1,7 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -19,9 +15,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-	"baseUrl": "./src"
+    "baseUrl": "./src"
   },
-  "include": [
-    "src"
-, "src/modules/TextBookPage/components/WordCard/SentencesBlock/.tsx"  ]
+  "include": ["src"]
 }
